+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     it will result in inlining the comments for those types everywhere they are referenced, #2303.
 -   Add `notRenderedTags` option. This option is similar to the `excludeTags` option, but while `excludeTags` will result in the
     tag being completely removed from the documentation, `notRenderedTags` only prevents it from being included when rendering.
+-   Changed the default `kindSortOrder` to put references last.
+-   Changed the default `sort` order to use `alphabetical-ignoring-documents` instead of `alphabetical`.
 
 TODO: